refactored contractwizard directories
